# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API for managing users",
        "version": "1.0.0"
    },
    "paths": {
        "/api/users/{id}/deposit": {
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Пополнение баланса пользователя",
                "operationId": "dbae60b9913e5392af4ff556593d63b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное пополнение",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    },
                    "400": {
                        "description": "Некорректные данные"
                    }
                }
            }
        },
        "/api/users/{id}/transfer": {
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Перевод средств между пользователями",
                "operationId": "97676797c03510c662d6d59927cfaddb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID отправителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_id",
                                    "amount"
                                ],
                                "properties": {
                                    "recipient_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Перевод успешно выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transfer successful"
                                        },
                                        "sender": {
                                            "$ref": "#/components/schemas/UserResource"
                                        },
                                        "recipient": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка перевода средств"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Получение всех пользователей с пагинацией",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Количество пользователей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/UserResource"
                                            }
                                        },
                                        "name": {
                                            "description": "John Doe",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "john@example.com",
                                            "type": "string"
                                        },
                                        "age": {
                                            "description": "30",
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "description": "1000",
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера"
                    }
                }
            },
            "post": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Создание нового пользователя",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "John Doe",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "john@example.com",
                                        "type": "string"
                                    },
                                    "age": {
                                        "description": "30",
                                        "type": "integer"
                                    },
                                    "balance": {
                                        "description": "1000",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "put": {
                "tags": [
                    "Пользователи"
                ],
                "summary": "Обновить пользователя",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "age"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешное обновление",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "UserResource": {
                "title": "User Resource",
                "description": "Ресурс пользователя, представляющий данные пользователя",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "age": {
                        "type": "integer",
                        "example": 30
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Пользователь",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "age": {
                        "type": "integer",
                        "example": 30
                    },
                    "balance": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Пользователи",
            "description": "Пользователи"
        }
    ]
}